pyproject.toml
[project]
name = "fastapi-security-jwt"
version = "0.3.0"
description = "FastAPI JWT authorization helpers: scope validation, token decoding, and JWK key fetching integration."
readme = "README.md"
authors = [{ name = "Connor Basile", email = "cbasile2@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "pyjwt[crypto]>=2.10.1",
    "pyjwt-key-fetcher>=0.8.0",
    "cachetools>=5.5.2",
]

[tool.pytest.ini_options]
anyio_mode = "auto"
addopts = "--cov --cov-report=term-missing --cov-report=xml --junitxml=junit.xml -o junit_family=legacy"

[tool.coverage.run]
source = ["src/fastapi_security_jwt"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py313" # or your Python version

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "D",   # pydocstyle
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modern Python syntax)
    "B",   # flake8-bugbear (catch common bugs)
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez (timezone-aware datetimes)
    "T10", # flake8-debugger (no debugger statements)
    "EM",  # flake8-errmsg (better error messages)
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie (misc linting)
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "PL",  # pylint
    "RUF", # ruff-specific rules
]

ignore = [
    "E501",    # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic values in comparisons
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"] # allow asserts in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Security"]

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "uvicorn>=0.40.0",
]
test = ["httpx>=0.28.1", "pytest>=9.0.2", "pytest-cov>=5.0.0"]
